add basic framework for the UpdatePacket handling
git-svn-id: file:///desktop/SourceForge%20Backup/empyrean/svnrepos/trunk@94 fea1f14b-4e95-431d-9825-d582385f3f73

+++ empyrean/src/base/WriterThread.cpp
@@ -40,7 +40,7 @@ namespace pyr {
             p->serialize(out);
             
             u16 id = PR_htons(p->getID());
-            u16 size = PR_htons(out.getSize());
+            u16 size = PR_htons(u16(out.getSize()));
             
             if (_socket->write(&id, sizeof(id)) != sizeof(id) ||
                 _socket->write(&size, sizeof(size)) != sizeof(size) ||
